fix(windows): remove Explorer preview handler
