+++ .github/workflows/release-windows.yaml
@@ -118,20 +118,6 @@ jobs:
         with:
           global-json-file: clients/windows/native/global.json
 
-      # Kept out of the job env: dotnet also honors MSBUILD_EXE_PATH and
-      # would resolve SDKs through the Visual Studio MSBuild.
-      - name: Locate MSBuild
-        id: msbuild
-        shell: pwsh
-        run: |
-          $vswhere = "${env:ProgramFiles(x86)}\Microsoft Visual Studio\Installer\vswhere.exe"
-          $msbuild = & $vswhere -latest -requires Microsoft.Component.MSBuild -find MSBuild\**\Bin\MSBuild.exe | Select-Object -First 1
-          if (-not $msbuild) { throw "MSBuild not found" }
-          "path=$msbuild" | Out-File -FilePath $env:GITHUB_OUTPUT -Append -Encoding utf8
-
-      - name: Integrate vcpkg with MSBuild
-        run: vcpkg integrate install
-
       - name: Install dependencies
         run: bun install --frozen-lockfile
 
@@ -148,11 +134,6 @@ jobs:
 
       - name: Build native helper
         run: bun scripts/build-native-helper.ts --arch ${{ matrix.arch }}
-
-      - name: Build preview handler
-        env:
-          MSBUILD_EXE_PATH: ${{ steps.msbuild.outputs.path }}
-        run: bun scripts/build-preview-handler.ts --arch ${{ matrix.arch }}
 
       - name: Build packaged runtime
         env:
@@ -191,7 +172,7 @@ jobs:
         run: bunx electron-builder --win --config electron-builder.config.cjs --publish never
 
       # Every binary in the signing manifest (app executable, CLI runtime,
-      # helper, preview-handler DLL) plus the installer must carry a valid
+      # helper) plus the installer must carry a valid
       # Authenticode signature from one certificate.
       - name: Verify signatures
         shell: pwsh
