--- clients/windows/native/Vellum.PreviewHandler/vcpkg.json
@@ -1,10 +1,0 @@
-{
-  "name": "vellum-preview-handler",
-  "version-string": "0.1.0",
-  "builtin-baseline": "a51bb4d1434e6d0927ff79db8033bed8522b85df",
-  "dependencies": ["nlohmann-json", "zlib"],
-  "overrides": [
-    { "name": "nlohmann-json", "version": "3.12.0", "port-version": 2 },
-    { "name": "zlib", "version": "1.3.2", "port-version": 1 }
-  ]
-}
