--- clients/windows/native/fixtures/vellum-bundle-contract/README.md
@@ -1,5 +1,0 @@
-# Vellum bundle contract fixtures
-
-`valid.vellum` uses the format version 2 manifest and DEFLATE settings from
-`assistant/src/bundler/app-bundler.ts`. Other fixtures cover malformed ZIP data,
-traversal, oversized icons, and manifest failures to detect format drift.
