--- clients/windows/native/fixtures/vellum-bundle-contract/malformed.vellum
@@ -1,1 +1,0 @@
-not a zip archive
