+++ clients/windows/package.json
@@ -15,15 +15,13 @@
     "dev:electron": "wait-on --timeout 30000 http://localhost:5173 && electron-vite dev",
     "build": "electron-vite build",
     "build:runtime": "bun run build:web && bun scripts/package-runtime.ts",
-    "build:preview-handler": "bun scripts/build-preview-handler.ts",
     "build:native-helper": "bun scripts/build-native-helper.ts",
     "typecheck": "bunx tsc --noEmit",
     "test": "bun scripts/run-tests.ts",
-    "test:ci": "bun run --filter=@vellumai/electron-desktop test && bun run --filter=@vellumai/native-sidecar test && bun scripts/run-tests.ts && bun scripts/build-native-helper.ts --check && bun scripts/build-preview-handler.ts --test-architecture && bun scripts/build-preview-handler.ts --native-test",
+    "test:ci": "bun run --filter=@vellumai/electron-desktop test && bun run --filter=@vellumai/native-sidecar test && bun scripts/run-tests.ts && bun scripts/build-native-helper.ts --check",
     "test:native-helper": "bun scripts/build-native-helper.ts --test",
-    "test:preview-handler": "bun scripts/build-preview-handler.ts --test-architecture && bun scripts/build-preview-handler.ts --native-test",
     "build:web": "cd ../web && bun install && bun run openapi-ts && bun run build && cd ../windows && mkdir -p resources && rm -rf resources/web-dist && cp -R ../web/dist resources/web-dist",
-    "pack": "bun run build:runtime && bun run build:native-helper && bun run build:preview-handler && electron-builder --win --config electron-builder.config.cjs --publish never",
+    "pack": "bun run build:runtime && bun run build:native-helper && electron-builder --win --config electron-builder.config.cjs --publish never",
     "pack:debug": "VELLUM_ENABLE_CHROME_DEVTOOLS=true bun run pack"
   },
   "devDependencies": {
