+++ clients/windows/scripts/package-smoke.ts
@@ -90,7 +90,6 @@ const main = async (): Promise<void> => {
     ["cli-runtime/bun.exe", "Bun runtime"],
     ["cli-runtime/runtime.json", "Runtime manifest"],
     [`native-helper/${arch}/Vellum.WindowsHelper.exe`, "Native helper"],
-    ["preview-handler/Vellum.PreviewHandler.dll", "Preview handler DLL"],
     ["tray.ico", "Tray icon"],
   ] as const) {
     assertExists(path.join(resources, relative), label);
